+++ src/jackdaw/serdes/resolver.clj
@@ -40,6 +40,7 @@
 
   Options:
   schema-registry-url - The URL for the schema registry
+  schema-registry-config - A map containing additional schema registry client configuration
   type-registry - A mapping per jackdaw.serdes.avro/+base-schema-type-registry+>
   read-only - Specifies that you will not be using the resulting serializer,
               and does not require a schema or schema-filename
@@ -48,7 +49,8 @@
   only the schema registry URL is required."
 
   [& options]
-  (let [{:keys [type-registry schema-registry-url schema-registry-client]}
+  (let [{:keys [type-registry schema-registry-url
+                schema-registry-client schema-registry-config]}
         (apply hash-map options)]
 
     (fn [{:keys [serde-keyword schema schema-filename key? read-only?] :as serde-config}]
@@ -67,5 +69,6 @@
                             (s/explain-data :jackdaw.serde/confluent-avro-serde serde-config)))
             (serde-fn schema-registry-url schema key? {:type-registry type-registry
                                                        :schema-registry-client schema-registry-client
+                                                       :schema-registry-config schema-registry-config
                                                        :read-only? read-only?}))
           (serde-fn))))))